+++ pkg/ansible/controller/reconcile_test.go
@@ -487,6 +487,7 @@ func TestReconcile(t *testing.T) {
 				GVK:             tc.GVK,
 				Runner:          tc.Runner,
 				Client:          tc.Client,
+				APIReader:       tc.Client,
 				EventHandlers:   tc.EventHandlers,
 				ReconcilePeriod: tc.ReconcilePeriod,
 				ManageStatus:    tc.ManageStatus,
@@ -501,10 +502,13 @@ func TestReconcile(t *testing.T) {
 			if tc.ExpectedObject != nil {
 				actualObject := &unstructured.Unstructured{}
 				actualObject.SetGroupVersionKind(tc.ExpectedObject.GroupVersionKind())
-				tc.Client.Get(context.TODO(), types.NamespacedName{
+				err := tc.Client.Get(context.TODO(), types.NamespacedName{
 					Name:      tc.ExpectedObject.GetName(),
 					Namespace: tc.ExpectedObject.GetNamespace(),
 				}, actualObject)
+				if err != nil {
+					t.Fatalf("Failed to get object: (%v)", err)
+				}
 				if !reflect.DeepEqual(actualObject.GetAnnotations(), tc.ExpectedObject.GetAnnotations()) {
 					t.Fatalf("Annotations are not the same\nexpected: %v\nactual: %v", tc.ExpectedObject.GetAnnotations(), actualObject.GetAnnotations())
 				}