+++ pkg/ansible/proxy/proxy.go
@@ -28,29 +28,24 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"strings"
-	"sync"
 
+	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/controllermap"
+	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/kubeconfig"
 	k8sRequest "github.com/operator-framework/operator-sdk/pkg/ansible/proxy/requestfactory"
+	osdkHandler "github.com/operator-framework/operator-sdk/pkg/handler"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metainternalversion "k8s.io/apimachinery/pkg/apis/meta/internalversion"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
-
-// ControllerMap - map of GVK to controller
-type ControllerMap struct {
-	sync.RWMutex
-	internal map[schema.GroupVersionKind]controller.Controller
-	watch    map[schema.GroupVersionKind]bool
-}
 
 type marshaler interface {
 	MarshalJSON() ([]byte, error)
@@ -60,7 +55,7 @@ type marshaler interface {
 // resource exists in our cache. If it does then there is no need to bombard
 // the APIserver with our request and we should write the response from the
 // proxy.
-func CacheResponseHandler(h http.Handler, informerCache cache.Cache, restMapper meta.RESTMapper, watchedNamespaces map[string]interface{}) http.Handler {
+func CacheResponseHandler(h http.Handler, informerCache cache.Cache, restMapper meta.RESTMapper, watchedNamespaces map[string]interface{}, cMap *controllermap.ControllerMap, injectOwnerRef bool) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 		switch req.Method {
 		case http.MethodGet:
@@ -143,17 +138,22 @@ func CacheResponseHandler(h http.Handler, informerCache cache.Cache, restMapper 
 				}
 				m = &un
 			} else {
-				un := unstructured.Unstructured{}
+				un := &unstructured.Unstructured{}
 				un.SetGroupVersionKind(k)
 				obj := client.ObjectKey{Namespace: r.Namespace, Name: r.Name}
-				err = informerCache.Get(context.Background(), obj, &un)
+				err = informerCache.Get(context.Background(), obj, un)
 				if err != nil {
 					// break here in case resource doesn't exist in cache but exists on APIserver
 					// This is very unlikely but provides user with expected 404
 					log.Info(fmt.Sprintf("Cache miss: %v, %v", k, obj))
 					break
 				}
-				m = &un
+				m = un
+				// Once we get the resource, we are going to attempt to recover the dependent watches here,
+				// This will happen in the background, and log errors.
+				if injectOwnerRef {
+					go recoverDependentWatches(req, un, cMap, restMapper)
+				}
 			}
 
 			i := bytes.Buffer{}
@@ -169,7 +169,9 @@ func CacheResponseHandler(h http.Handler, informerCache cache.Cache, restMapper 
 			w.Header().Set("Content-Type", "application/json")
 			// Set X-Cache header to signal that response is served from Cache
 			w.Header().Set("X-Cache", "HIT")
-			json.Indent(&i, resp, "", "  ")
+			if err := json.Indent(&i, resp, "", "  "); err != nil {
+				log.Error(err, "Failed to indent json")
+			}
 			_, err = w.Write(i.Bytes())
 			if err != nil {
 				log.Error(err, "Failed to write response")
@@ -184,10 +186,42 @@ func CacheResponseHandler(h http.Handler, informerCache cache.Cache, restMapper 
 	})
 }
 
+func recoverDependentWatches(req *http.Request, un *unstructured.Unstructured, cMap *controllermap.ControllerMap, restMapper meta.RESTMapper) {
+	ownerRef, err := getRequestOwnerRef(req)
+	if err != nil {
+		log.Error(err, "Could not get ownerRef from proxy")
+		return
+	}
+
+	for _, oRef := range un.GetOwnerReferences() {
+		if oRef.APIVersion == ownerRef.APIVersion && oRef.Kind == ownerRef.Kind {
+			err := addWatchToController(ownerRef, cMap, un, restMapper, true)
+			if err != nil {
+				log.Error(err, "Could not recover dependent resource watch", "owner", ownerRef)
+				return
+			}
+		}
+	}
+	if typeString, ok := un.GetAnnotations()[osdkHandler.TypeAnnotation]; ok {
+		ownerGV, err := schema.ParseGroupVersion(ownerRef.APIVersion)
+		if err != nil {
+			log.Error(err, "Could not get ownerRef from proxy")
+			return
+		}
+		if typeString == fmt.Sprintf("%v.%v", ownerRef.Kind, ownerGV.Group) {
+			err := addWatchToController(ownerRef, cMap, un, restMapper, false)
+			if err != nil {
+				log.Error(err, "Could not recover dependent resource watch", "owner", ownerRef)
+				return
+			}
+		}
+	}
+}
+
 // InjectOwnerReferenceHandler will handle proxied requests and inject the
 // owner refernece found in the authorization header. The Authorization is
 // then deleted so that the proxy can re-set with the correct authorization.
-func InjectOwnerReferenceHandler(h http.Handler, cMap *ControllerMap, restMapper meta.RESTMapper) http.Handler {
+func InjectOwnerReferenceHandler(h http.Handler, cMap *controllermap.ControllerMap, restMapper meta.RESTMapper, watchedNamespaces map[string]interface{}) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 		switch req.Method {
 		case http.MethodPost:
@@ -206,24 +240,13 @@ func InjectOwnerReferenceHandler(h http.Handler, cMap *ControllerMap, restMapper
 				break
 			}
 			log.Info("Injecting owner reference")
-
-			user, _, ok := req.BasicAuth()
-			if !ok {
-				log.Error(errors.New("basic auth header not found"), "")
-				w.Header().Set("WWW-Authenticate", "Basic realm=\"Operator Proxy\"")
-				http.Error(w, "", http.StatusUnauthorized)
-				return
-			}
-			authString, err := base64.StdEncoding.DecodeString(user)
+			owner, err := getRequestOwnerRef(req)
 			if err != nil {
-				m := "Could not base64 decode username"
+				m := "Could not get owner reference"
 				log.Error(err, m)
-				http.Error(w, m, http.StatusBadRequest)
+				http.Error(w, m, http.StatusInternalServerError)
 				return
 			}
-			owner := metav1.OwnerReference{}
-			json.Unmarshal(authString, &owner)
-			log.Info(fmt.Sprintf("Owner: %#v", owner))
 
 			body, err := ioutil.ReadAll(req.Body)
 			if err != nil {
@@ -240,7 +263,33 @@ func InjectOwnerReferenceHandler(h http.Handler, cMap *ControllerMap, restMapper
 				http.Error(w, m, http.StatusBadRequest)
 				return
 			}
-			data.SetOwnerReferences(append(data.GetOwnerReferences(), owner))
+
+			addOwnerRef, err := shouldAddOwnerRef(data, owner, restMapper)
+			if err != nil {
+				m := "Could not determine if we should add owner ref"
+				log.Error(err, m)
+				http.Error(w, m, http.StatusBadRequest)
+				return
+			}
+			if addOwnerRef {
+				data.SetOwnerReferences(append(data.GetOwnerReferences(), owner.OwnerReference))
+			} else {
+				ownerGV, err := schema.ParseGroupVersion(owner.APIVersion)
+				if err != nil {
+					m := fmt.Sprintf("could not get broup version for: %v", owner)
+					log.Error(err, m)
+					http.Error(w, m, http.StatusBadRequest)
+					return
+				}
+				a := data.GetAnnotations()
+				if a == nil {
+					a = map[string]string{}
+				}
+				a[osdkHandler.NamespacedNameAnnotation] = strings.Join([]string{owner.Namespace, owner.Name}, "/")
+				a[osdkHandler.TypeAnnotation] = fmt.Sprintf("%v.%v", owner.Kind, ownerGV.Group)
+
+				data.SetAnnotations(a)
+			}
 			newBody, err := json.Marshal(data.Object)
 			if err != nil {
 				m := "Could not serialize body"
@@ -251,30 +300,17 @@ func InjectOwnerReferenceHandler(h http.Handler, cMap *ControllerMap, restMapper
 			log.V(1).Info("Serialized body", "Body", string(newBody))
 			req.Body = ioutil.NopCloser(bytes.NewBuffer(newBody))
 			req.ContentLength = int64(len(newBody))
-			dataMapping, err := restMapper.RESTMapping(data.GroupVersionKind().GroupKind(), data.GroupVersionKind().Version)
-			if err != nil {
-				m := fmt.Sprintf("Could not get rest mapping for: %v", data.GroupVersionKind())
-				log.Error(err, m)
-				http.Error(w, m, http.StatusInternalServerError)
-				return
-			}
-			// We need to determine whether or not the owner is a cluster-scoped
-			// resource because enqueue based on an owner reference does not work if
-			// a namespaced resource owns a cluster-scoped resource
-			ownerGV, err := schema.ParseGroupVersion(owner.APIVersion)
-			ownerMapping, err := restMapper.RESTMapping(schema.GroupKind{Kind: owner.Kind, Group: ownerGV.Group}, ownerGV.Version)
-			if err != nil {
-				m := fmt.Sprintf("could not get rest mapping for: %v", owner)
-				log.Error(err, m)
-				http.Error(w, m, http.StatusInternalServerError)
-				return
-			}
 
-			dataClusterScoped := dataMapping.Scope.Name() != meta.RESTScopeNameRoot
-			ownerClusterScoped := ownerMapping.Scope.Name() != meta.RESTScopeNameRoot
-			if !ownerClusterScoped || dataClusterScoped {
-				// add watch for resource
-				err = addWatchToController(owner, cMap, data)
+			// add watch for resource
+			// check if resource doesn't exist in watched namespaces
+			// if watchedNamespaces[""] exists then we are watching all namespaces
+			// and want to continue
+			// This is making sure we are not attempting to watch a resource outside of the
+			// namespaces that the cache can watch.
+			_, allNsPresent := watchedNamespaces[metav1.NamespaceAll]
+			_, reqNsPresent := watchedNamespaces[r.Namespace]
+			if allNsPresent || reqNsPresent {
+				err = addWatchToController(owner, cMap, data, restMapper, addOwnerRef)
 				if err != nil {
 					m := "could not add watch to controller"
 					log.Error(err, m)
@@ -289,6 +325,40 @@ func InjectOwnerReferenceHandler(h http.Handler, cMap *ControllerMap, restMapper
 	})
 }
 
+func shouldAddOwnerRef(data *unstructured.Unstructured, owner kubeconfig.NamespacedOwnerReference, restMapper meta.RESTMapper) (bool, error) {
+	dataMapping, err := restMapper.RESTMapping(data.GroupVersionKind().GroupKind(), data.GroupVersionKind().Version)
+	if err != nil {
+		m := fmt.Sprintf("Could not get rest mapping for: %v", data.GroupVersionKind())
+		log.Error(err, m)
+		return false, err
+
+	}
+	// We need to determine whether or not the owner is a cluster-scoped
+	// resource because enqueue based on an owner reference does not work if
+	// a namespaced resource owns a cluster-scoped resource
+	ownerGV, err := schema.ParseGroupVersion(owner.APIVersion)
+	if err != nil {
+		m := fmt.Sprintf("could not get group version for: %v", owner)
+		log.Error(err, m)
+		return false, err
+	}
+	ownerMapping, err := restMapper.RESTMapping(schema.GroupKind{Kind: owner.Kind, Group: ownerGV.Group}, ownerGV.Version)
+	if err != nil {
+		m := fmt.Sprintf("could not get rest mapping for: %v", owner)
+		log.Error(err, m)
+		return false, err
+	}
+
+	dataNamespaceScoped := dataMapping.Scope.Name() != meta.RESTScopeNameRoot
+	ownerNamespaceScoped := ownerMapping.Scope.Name() != meta.RESTScopeNameRoot
+
+	if dataNamespaceScoped && ownerNamespaceScoped && data.GetNamespace() == owner.Namespace {
+		return true, nil
+	}
+	return false, nil
+}
+
+// RequestLogHandler - log the requests that come through the proxy.
 func RequestLogHandler(h http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 		// read body
@@ -321,7 +391,7 @@ type Options struct {
 	KubeConfig        *rest.Config
 	Cache             cache.Cache
 	RESTMapper        meta.RESTMapper
-	ControllerMap     *ControllerMap
+	ControllerMap     *controllermap.ControllerMap
 	WatchedNamespaces []string
 	DisableCache      bool
 }
@@ -359,7 +429,9 @@ func Run(done chan error, o Options) error {
 		}
 		stop := make(chan struct{})
 		go func() {
-			informerCache.Start(stop)
+			if err := informerCache.Start(stop); err != nil {
+				log.Error(err, "Failed to start informer cache")
+			}
 			defer close(stop)
 		}()
 		log.Info("Waiting for cache to sync...")
@@ -372,13 +444,13 @@ func Run(done chan error, o Options) error {
 	}
 
 	if !o.NoOwnerInjection {
-		server.Handler = InjectOwnerReferenceHandler(server.Handler, o.ControllerMap, o.RESTMapper)
+		server.Handler = InjectOwnerReferenceHandler(server.Handler, o.ControllerMap, o.RESTMapper, watchedNamespaceMap)
 	}
 	if o.LogRequests {
 		server.Handler = RequestLogHandler(server.Handler)
 	}
 	if !o.DisableCache {
-		server.Handler = CacheResponseHandler(server.Handler, o.Cache, o.RESTMapper, watchedNamespaceMap)
+		server.Handler = CacheResponseHandler(server.Handler, o.Cache, o.RESTMapper, watchedNamespaceMap, o.ControllerMap, !o.NoOwnerInjection)
 	}
 
 	l, err := server.Listen(o.Address, o.Port)
@@ -392,70 +464,100 @@ func Run(done chan error, o Options) error {
 	return nil
 }
 
-func addWatchToController(owner metav1.OwnerReference, cMap *ControllerMap, resource *unstructured.Unstructured) error {
-	gv, err := schema.ParseGroupVersion(owner.APIVersion)
+func addWatchToController(owner kubeconfig.NamespacedOwnerReference, cMap *controllermap.ControllerMap, resource *unstructured.Unstructured, restMapper meta.RESTMapper, useOwnerRef bool) error {
+	dataMapping, err := restMapper.RESTMapping(resource.GroupVersionKind().GroupKind(), resource.GroupVersionKind().Version)
 	if err != nil {
+		m := fmt.Sprintf("Could not get rest mapping for: %v", resource.GroupVersionKind())
+		log.Error(err, m)
+		return err
+
+	}
+	ownerGV, err := schema.ParseGroupVersion(owner.APIVersion)
+	if err != nil {
+		m := fmt.Sprintf("could not get broup version for: %v", owner)
+		log.Error(err, m)
 		return err
 	}
-	gvk := schema.GroupVersionKind{
-		Group:   gv.Group,
-		Version: gv.Version,
-		Kind:    owner.Kind,
+	ownerMapping, err := restMapper.RESTMapping(schema.GroupKind{Kind: owner.Kind, Group: ownerGV.Group}, ownerGV.Version)
+	if err != nil {
+		m := fmt.Sprintf("could not get rest mapping for: %v", owner)
+		log.Error(err, m)
+		return err
 	}
-	c, watch, ok := cMap.Get(gvk)
+
+	dataNamespaceScoped := dataMapping.Scope.Name() != meta.RESTScopeNameRoot
+	contents, ok := cMap.Get(ownerMapping.GroupVersionKind)
 	if !ok {
 		return errors.New("failed to find controller in map")
 	}
+	wMap := contents.WatchMap
+	uMap := contents.UIDMap
 	u := &unstructured.Unstructured{}
-	u.SetGroupVersionKind(gvk)
+	u.SetGroupVersionKind(ownerMapping.GroupVersionKind)
 	// Add a watch to controller
-	if watch {
-		log.Info("Watching child resource", "kind", resource.GroupVersionKind(), "enqueue_kind", u.GroupVersionKind())
-		err = c.Watch(&source.Kind{Type: resource}, &handler.EnqueueRequestForOwner{OwnerType: u})
-		if err != nil {
-			return err
+	if contents.WatchDependentResources {
+		// Store UID
+		uMap.Store(owner.UID, types.NamespacedName{
+			Name:      owner.Name,
+			Namespace: owner.Namespace,
+		})
+		_, exists := wMap.Get(resource.GroupVersionKind())
+		// If already watching resource no need to add a new watch
+		if exists {
+			return nil
+		}
+		// Store watch in map
+		wMap.Store(resource.GroupVersionKind())
+		// Use EnqueueRequestForOwner unless user has configured watching cluster scoped resources and we have to
+		switch {
+		case useOwnerRef:
+			log.Info("Watching child resource", "kind", resource.GroupVersionKind(), "enqueue_kind", u.GroupVersionKind())
+			// Store watch in map
+			err := contents.Controller.Watch(&source.Kind{Type: resource}, &handler.EnqueueRequestForOwner{OwnerType: u})
+			if err != nil {
+				return err
+			}
+		case !useOwnerRef && dataNamespaceScoped:
+			typeString := fmt.Sprintf("%v.%v", owner.Kind, ownerGV.Group)
+			log.Info("Watching child resource", "kind", resource.GroupVersionKind(), "enqueue_annotation_type", typeString)
+			err = contents.Controller.Watch(&source.Kind{Type: resource}, &osdkHandler.EnqueueRequestForAnnotation{Type: typeString})
+			if err != nil {
+				return err
+			}
+		case contents.WatchClusterScopedResources:
+			typeString := fmt.Sprintf("%v.%v", owner.Kind, ownerGV.Group)
+			log.Info("Watching child resource which can be cluster-scoped", "kind", resource.GroupVersionKind(), "enqueue_annotation_type", typeString)
+			// Add watch
+			err = contents.Controller.Watch(&source.Kind{Type: resource}, &osdkHandler.EnqueueRequestForAnnotation{Type: typeString})
+			if err != nil {
+				return err
+			}
 		}
 	}
 	return nil
 }
 
-// NewControllerMap returns a new object that contains a mapping between GVK
-// and controller
-func NewControllerMap() *ControllerMap {
-	return &ControllerMap{
-		internal: make(map[schema.GroupVersionKind]controller.Controller),
-		watch:    make(map[schema.GroupVersionKind]bool),
-	}
-}
-
-// Get - Returns a controller given a GVK as the key. `watch` in the return
-// specifies whether or not the operator will watch dependent resources for
-// this controller. `ok` returns whether the query was successful. `controller`
-// is the associated controller-runtime controller object.
-func (cm *ControllerMap) Get(key schema.GroupVersionKind) (controller controller.Controller, watch, ok bool) {
-	cm.RLock()
-	defer cm.RUnlock()
-	result, ok := cm.internal[key]
+func getRequestOwnerRef(req *http.Request) (kubeconfig.NamespacedOwnerReference, error) {
+	owner := kubeconfig.NamespacedOwnerReference{}
+	user, _, ok := req.BasicAuth()
 	if !ok {
-		return result, false, ok
+		return owner, errors.New("basic auth header not found")
 	}
-	watch, ok = cm.watch[key]
-	return result, watch, ok
-}
-
-// Delete - Deletes associated GVK to controller mapping from the ControllerMap
-func (cm *ControllerMap) Delete(key schema.GroupVersionKind) {
-	cm.Lock()
-	defer cm.Unlock()
-	delete(cm.internal, key)
-}
-
-// Store - Adds a new GVK to controller mapping. Also creates a mapping between
-// GVK and a boolean `watch` that specifies whether this controller is watching
-// dependent resources.
-func (cm *ControllerMap) Store(key schema.GroupVersionKind, value controller.Controller, watch bool) {
-	cm.Lock()
-	defer cm.Unlock()
-	cm.internal[key] = value
-	cm.watch[key] = watch
+	authString, err := base64.StdEncoding.DecodeString(user)
+	if err != nil {
+		m := "Could not base64 decode username"
+		log.Error(err, m)
+		return owner, err
+	}
+	// Set owner to NamespacedOwnerReference, which has metav1.OwnerReference
+	// as a subset along with the Namespace of the owner. Please see the
+	// kubeconfig.NamespacedOwnerReference type for more information. The
+	// namespace is required when creating the reconcile requests.
+	json.Unmarshal(authString, &owner)
+	if err := json.Unmarshal(authString, &owner); err != nil {
+		m := "Could not unmarshal auth string"
+		log.Error(err, m)
+		return owner, err
+	}
+	return owner, err
 }