+++ pkg/ansible/proxy/proxy_test.go
@@ -21,6 +21,9 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/operator-framework/operator-sdk/internal/util/fileutil"
+	"github.com/operator-framework/operator-sdk/pkg/ansible/proxy/controllermap"
+
 	kcorev1 "k8s.io/api/core/v1"
 	kmetav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -36,7 +39,7 @@ func TestHandler(t *testing.T) {
 		t.Fatalf("Failed to instantiate manager: %v", err)
 	}
 	done := make(chan error)
-	cMap := NewControllerMap()
+	cMap := controllermap.NewControllerMap()
 	err = Run(done, Options{
 		Address:           "localhost",
 		Port:              8888,
@@ -56,7 +59,11 @@ func TestHandler(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Error getting pod from proxy: %v", err)
 	}
-	defer resp.Body.Close()
+	defer func() {
+		if err := resp.Body.Close(); err != nil && !fileutil.IsClosedError(err) {
+			t.Errorf("Failed to close response body: (%v)", err)
+		}
+	}()
 	body, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
 		t.Fatalf("Error reading response body: %v", err)